refactor: globalize Footer and fix AuthLayout gap
