+++ client/src/modules/analytics/TutorAnalyticsDashboard.jsx
@@ -5,7 +5,7 @@ import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { apiRequest } from "../../services/apiClient.js";
 import Navbar from "../../shared/components/Navbar";
-import Footer from "../../modules/landing/components/Footer";
+import Footer from "../../shared/components/Footer";
 
 import { useDocumentTitle } from "../../hooks/useDocumentTitle";
