+++ client/src/modules/classrooms/pages/ClassroomsDashboard.jsx
@@ -9,7 +9,7 @@ import {
   getActiveClassroomSessions,
 } from "../services/classroomService";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
