+++ client/src/modules/dashboard/pages/CoverLetterHistoryPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import { getCoverLetterHistory } from "../services/dashboardService";
 import {
