+++ client/src/modules/landing/pages/BlogPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../components/Footer";
+import Footer from "../../../shared/components/Footer";
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
 
 const BlogPage = () => {
