+++ client/src/modules/landing/pages/CareersPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../components/Footer";
+import Footer from "../../../shared/components/Footer";
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
 import { ArrowLeft, MapPin, Clock, ArrowRight } from "lucide-react";
