+++ client/src/modules/mock-interview/pages/InterviewResults.jsx
@@ -18,7 +18,7 @@ import {
   ArrowLeft,
 } from "lucide-react";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
 import logger from "../../../utils/logger";
