+++ client/src/modules/mock-interview/pages/TutorInterviewsList.jsx
@@ -4,7 +4,7 @@ import { useSelector } from "react-redux";
 import { Clock, CheckCircle, Video, ArrowRight, User, ArrowLeft } from "lucide-react";
 import { apiRequest } from "../../../services/apiClient.js";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import { useDocumentTitle } from "../../../hooks/useDocumentTitle";
