+++ client/src/modules/notifications/pages/NotificationsPage.jsx
@@ -9,7 +9,7 @@ import {
   Filter,
 } from "lucide-react";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import NotificationCard from "../components/NotificationCard";
 import useNotifications from "../hooks/useNotifications";
