+++ client/src/modules/recruiter-jobs/pages/CreateJobPostingPage.jsx
@@ -3,7 +3,7 @@ import { useSelector } from "react-redux";
 import { useNavigate, Link } from "react-router-dom";
 import { ArrowLeft, Briefcase } from "lucide-react";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import JobPostingForm from "../components/JobPostingForm";
 import { createJobPosting } from "../services/jobPostingService";
