+++ client/src/modules/recruiter-jobs/pages/RecruiterApplicantsPage.jsx
@@ -25,7 +25,7 @@ import {
   KanbanSquare
 } from 'lucide-react';
 import Navbar from '../../../shared/components/Navbar';
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import { Button, LoadingState, ErrorState, EmptyState, StatusUpdateModal, StatusTimeline } from '../../../shared/components';
 import ApplicantsKanbanBoard from '../components/ApplicantsKanbanBoard';
