+++ client/src/modules/recruiter-jobs/pages/__tests__/RecruiterJobsPage.test.jsx
@@ -7,7 +7,7 @@ import { MemoryRouter } from 'react-router-dom'
 import RecruiterJobsPage from '../RecruiterJobsPage'
 import * as jobPostingService from '../../services/jobPostingService'
 import { ToastProvider } from '../../../../shared/components/toast/ToastProvider'
-import Footer from "../../../../modules/landing/components/Footer";
+import Footer from "../../../../shared/components/Footer";
 
 
 // Mock the service
