+++ client/src/modules/student-jobs/pages/MyApplicationsPage.jsx
@@ -16,7 +16,7 @@ import {
   List,
 } from "lucide-react";
 import Navbar from "../../../shared/components/Navbar";
-import Footer from "../../../modules/landing/components/Footer";
+import Footer from "../../../shared/components/Footer";
 
 import LoadingState from "../../../shared/components/LoadingState";
 import ConfirmDialog from "../../../shared/components/ConfirmDialog";
