+++ client/src/shared/components/Footer.jsx
